+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added ability to collapse post in post page
 - Added ability to change app language in settings
 - Added ability to show/hide read posts in user settings
+- Added post and comment previews to settings, and reorganized settings page
 
 ## 0.2.6 - 2023-11-22
 ### Fixed